+++ amplitude.js
@@ -139,7 +139,9 @@ var DEFAULT_OPTIONS = {
   saveEvents: true,
   sessionTimeout: 30 * 60 * 1000,
   unsentKey: 'amplitude_unsent',
-  uploadBatchSize: 100
+  uploadBatchSize: 100,
+  batchEvents: false,
+  eventUploadThreshold: 30
 };
 var LocalStorageKeys = {
   LAST_EVENT_ID: 'amplitude_lastEventId',
@@ -188,10 +190,14 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
       if (opt_config.includeReferrer !== undefined) {
         this.options.includeReferrer = !!opt_config.includeReferrer;
       }
+      if (opt_config.batchEvents !== undefined) {
+        this.options.batchEvents = !!opt_config.batchEvents;
+      }
       this.options.platform = opt_config.platform || this.options.platform;
       this.options.language = opt_config.language || this.options.language;
       this.options.sessionTimeout = opt_config.sessionTimeout || this.options.sessionTimeout;
       this.options.uploadBatchSize = opt_config.uploadBatchSize || this.options.uploadBatchSize;
+      this.options.eventUploadThreshold = opt_config.eventUploadThreshold || this.options.eventUploadThreshold;
       this.options.savedMaxCount = opt_config.savedMaxCount || this.options.savedMaxCount;
     }
 
@@ -222,7 +228,7 @@ Amplitude.prototype.init = function(apiKey, opt_userId, opt_config) {
         }
       }
     }
-    if (this._unsentEvents.length > 0) {
+    if (this.shouldSendEvents()) {
       this.sendEvents();
     }
 
@@ -253,6 +259,13 @@ Amplitude.prototype.isNewSession = function() {
 Amplitude.prototype.nextEventId = function() {
   this._eventId++;
   return this._eventId;
+};
+
+Amplitude.prototype.shouldSendEvents = function() {
+  var batchEvents = this.options.batchEvents;
+  var numEvents = this._unsentEvents.length;
+  var threshold = this.options.eventUploadThreshold;
+  return (!batchEvents && numEvents > 0) || (batchEvents && numEvents >= threshold);
 };
 
 var _loadCookieData = function(scope) {
@@ -477,7 +490,9 @@ Amplitude.prototype._logEvent = function(eventType, eventProperties, apiProperti
       this.saveEvents();
     }
 
-    this.sendEvents();
+    if (this.shouldSendEvents()){
+      this.sendEvents();
+    }
 
     return eventId;
   } catch (e) {
@@ -557,7 +572,7 @@ Amplitude.prototype.sendEvents = function() {
           }
 
           // Send more events if any queued during previous send.
-          if (scope._unsentEvents.length > 0) {
+          if (scope.shouldSendEvents()) {
             scope.sendEvents();
           }
         } else if (status === 413) {